requirements.txt
absl-py==2.3.1
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.4
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argbind==0.3.9
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
autopage==0.5.2
babel==2.17.0
beautifulsoup4==4.13.4
bitsandbytes==0.45.0
bleach==6.2.0
certifi==2025.7.9
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cliff==4.10.0
cmaes==0.12.0
cmd2==2.7.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.3
conformer==0.3.2
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.6
csvw==3.5.1
cycler==0.12.1
Cython==3.1.2
datasets==2.18.0
debugpy==1.8.16
decorator==5.2.1
deepspeed==0.17.2
defusedxml==0.7.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
diffusers==0.25.0
dill==0.3.8
distlib==0.4.0
dlinfo==2.0.0
docstring_parser==0.17.0
editdistance==0.8.1
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.2
ffmpy==0.6.0
filelock==3.18.0
fire==0.7.1
flatbuffers==25.2.10
flatten-dict==0.4.2
fonttools==4.58.5
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.2.0
funasr==1.2.6
future==1.0.0
gdown==5.1.0
gradio==3.43.2
gradio_client==0.5.0
greenlet==3.2.4
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.25.1
humanfriendly==10.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
HyperPyYAML==1.2.2
identify==2.6.13
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.3.1
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
julius==0.2.7
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaldiio==2.18.1
kiwisolver==1.4.8
language-tags==1.2.0
lark==1.2.2
latex2mathml==3.78.0
lazy_loader==0.4
librosa==0.10.2
lightning==2.2.4
lightning-utilities==0.14.3
# Editable install with no version control (llama_omni2==1.0.0)
-e /inspire/hdd/project/socialsimulation/chensiyuan-CZXS24210113/speechmed/SpeechMedAssist
llvmlite==0.44.0
loguru==0.7.3
loralib==0.1.2
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==2.1.5
matcha-tts @ file:///inspire/hdd/project/socialsimulation/chensiyuan-CZXS24210113/speechmed/SpeechMedAssist/CosyVoice/third_party/Matcha-TTS
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
modelscope==1.28.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
narwhals==2.1.2
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nodeenv==1.9.1
notebook==7.4.5
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
nvitop==1.5.3
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.18.0
openai-whisper==20231117
optuna==2.10.1
orjson==3.10.18
oss2==2.19.1
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pbr==7.0.0
peft==0.14.0
pexpect==4.9.0
phonemizer==3.3.0
pillow==10.4.0
piper-phonemize==1.1.0
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
pre_commit==4.3.0
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.19.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
pynini==2.1.5
pynndescent==0.5.13
pyparsing==3.2.3
pyperclip==1.9.0
pyrootutils==1.0.4
PySocks==1.7.1
pystoi==0.4.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytorch-wpe==0.0.1
pyttsx3==2.99
pytz==2025.2
pyworld==0.3.4
PyYAML==6.0.2
pyzmq==27.0.1
randomname==0.2.1
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.0.0
rich-argparse==1.7.1
rootutils==1.0.7
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.2
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
segments==2.3.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.46.2
stevedore==5.4.1
svgwrite==1.4.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==0.6.13
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.3.1
torch-complex==0.4.4
torch-stoi==0.2.3
torchaudio==2.3.1
torchmetrics==1.7.4
torchvision==0.18.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.43.4
triton==2.3.1
ttsfrd @ file:///inspire/hdd/project/socialsimulation/chensiyuan-CZXS24210113/speechmed/SpeechMedAssist/weight/CosyVoice-ttsfrd/ttsfrd-0.4.2-cp310-cp310-linux_x86_64.whl#sha256=149514a05c2a7a01076170d2923efd231fc579f8e48403ab523814c8256efe14
ttsfrd-dependency @ file:///inspire/hdd/project/socialsimulation/chensiyuan-CZXS24210113/speechmed/SpeechMedAssist/weight/CosyVoice-ttsfrd/ttsfrd_dependency-0.1-py3-none-any.whl#sha256=060a53f0650d12839983afdcfb052b049d7cf5c62344a00fee3a7344582aaf6f
typeguard==4.4.4
typer==0.16.0
types-python-dateutil==2.9.0.20250809
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
Unidecode==1.4.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.0
virtualenv==20.34.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
WeTextProcessing==1.0.3
wget==3.2
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
